move validation function to validation.py
