+++ src/python/impactx/dashboard/Input/lattice/ui.py
@@ -147,7 +147,7 @@ def process_if_variable(index, parameter_name, ui_input, parameter_type):
     is_variable, variable_index = LatticeVariableHandler.determine_if_existing_variable(
         var_name
     )
-    is_potential_variable = LatticeConfigurationHelper.is_valid_input_name(var_name)
+    is_potential_variable = DashboardValidation.is_valid_input_name(var_name)
 
     if is_variable:
         sim_value = state.variables[variable_index]["value"]
@@ -180,7 +180,7 @@ def on_lattice_element_parameter_change(
         state.lattice_elements_using_variables.pop(key, None)
 
     if parameter_name in ["name", "backend"]:
-        if not LatticeConfigurationHelper.is_valid_input_name(ui_input):
+        if not DashboardValidation.is_valid_input_name(ui_input):
             error_message = ["Must be a valid Python identifier"]
         else:
             error_message = []
